+++ src/main/groovy/life/qbic/datamodel/datasets/datastructure/files/nfcore/RunId.groovy
@@ -5,7 +5,7 @@ import life.qbic.datamodel.datasets.datastructure.files.DataFile
 /**
  A specialisation of a DataFile, represents a nf-core RunId file
  *
- * @since 2.5.0
+ * @since 2.6.0
  */
 class RunId extends DataFile {