+++ CHANGELOG.rst
@@ -5,6 +5,22 @@ Changelog
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 
+2.6.0 (2021-05-10)
+------------------
+
+**Added**
+
+* Add json schema and wrapper for imaging data ``life.qbic.datamodel.datasets.imaging.ImageMetadata`` (`#199 <https://github.com/qbicsoftware/data-model-lib/pull/199>`_)
+
+* Provide new classes that describe incoming nf-core bioinformatic pipeline data structures and contains it in a dedicated NfCorePipelineResult class (`#198 <https://github.com/qbicsoftware/data-model-lib/pull/198>`_)
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**
+
+
 2.5.0 (2021-04-26)
 ------------------